More non-Linux-related merging in ctp2_code/libs/anet.  The remaining diff between trunk and the Linux branch is all wicked and tricksy, more or less.
git-svn-id: http://ctp2.darkdust.net/anonsvn/branches/linux@772 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/libs/anet/src/mac/main/MW2TransportMain.cpp
@@ -35,7 +35,7 @@ void __myterminate(CFragInitBlockPtr ibp) {
 	
 	//__terminate();
 }
-void *operator new(size_t size) {
+void *operator new(unsigned long size) {
 	return NewPtrSys(size);
 }
 void operator delete(void *ptr) {
